Refactor code to be more consistent with the frontend
