+++ backend/src/functions/index.js
@@ -1,5 +1,20 @@
-export { default as createTodo } from './create-todo.js';
-export { default as deleteTodo } from './delete-todo.js';
-export { default as fetchTodos } from './fetch-todos.js';
-export { default as updateTodo } from './update-todo.js';
-export { default as throwIfEmpty } from './throw-if-empty.js';
+export { default as createTodo } from './createTodo.js';
+export { default as deleteTodo } from './deleteTodo.js';
+export { default as fetchTodos } from './fetchTodos.js';
+export { default as updateTodo } from './updateTodo.js';
+
+/**
+ * Throw an error if the supplied data is falsy, else return it
+ *
+ * @param {any}   data
+ * @param {Error} err
+ *
+ * @return {void}
+ */
+export function throwIfEmpty (data, err) {
+    if (!data) {
+        throw err;
+    }
+
+    return data;
+}
