+++ backend/src/index.js
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose';
 import { app } from './bootstrap/app.js';
-import { DATABASE_CONNECTION_URI, SERVICE_PORT } from './configs/index.js';
+import { DATABASE_CONNECTION_URI, SERVICE_PORT } from './config/index.js';
 
 /* Start the Express server */
 app.listen(SERVICE_PORT, async () => {
